Revert "Merge remote-tracking branch 'upstream/master' into 20240805_works"
This reverts commit bd9406d5463287cc5dd06ce8e968bbba2bc65033, reversing
changes made to 9ae5131b61dc1626a53f83fcef0732fd959bd708.

+++ src/utils/DbgHelpDyn.cpp
@@ -430,7 +430,7 @@ void GetThreadCallstack(str::Str& s, DWORD threadId) {
     }
 
     DWORD res = SuspendThread(hThread);
-    if ((DWORD)(-1) == res) {
+    if (-1 == res) {
         s.Append("Failed to SuspendThread()\n");
     } else {
         CONTEXT ctx{};
